+++ packages/vant/src/renderers/SelectorRenderer/types.ts
@@ -4,7 +4,7 @@
  * @module renderers/SelectorRenderer/types
  */
 
-import type { SchemxVueBaseComponentProps } from "@schemx/vue"
+import type { SchemxBaseComponentProps } from "@schemx/vue"
 
 export type SelectValue = string | number | (string | number)[]
 
@@ -57,10 +57,7 @@ export interface SelectorProps {
 export interface SelectorRendererProps
   extends
     SelectorProps,
-    Omit<
-      SchemxVueBaseComponentProps,
-      "onChange" | "onBlur" | "value" | "onUpdate:value"
-    > {
+    Omit<SchemxBaseComponentProps, "onChange" | "onBlur" | "value" | "onUpdate:value"> {
   /** 当前值 */
   value?: SelectValue
   /** 值变化回调 */