+++ packages/vant/src/renderers/StepperRenderer/types.ts
@@ -6,7 +6,7 @@
 
 import type { StepperProps } from "vant"
 
-import type { SchemxVueBaseComponentProps } from "@schemx/vue"
+import type { SchemxBaseComponentProps } from "@schemx/vue"
 
 export type StepperValue = StepperProps["modelValue"]
 
@@ -18,7 +18,7 @@ export type StepperValue = StepperProps["modelValue"]
 export interface StepperRendererProps
   /* @vue-ignore */
   extends
-    Omit<SchemxVueBaseComponentProps, "onChange" | "onBlur" | "value" | "onUpdate:value">,
+    Omit<SchemxBaseComponentProps, "onChange" | "onBlur" | "value" | "onUpdate:value">,
     /* @vue-ignore */
     Partial<Omit<StepperProps, "modelValue" | "onUpdate:modelValue" | "onChange">> {
   /** 当前值 */